requirements.txt
Appium-Python-Client>=3.1.1
pytest>=8.0.0
pytest-xdist>=3.5.0
selenium>=4.16.0
python-dotenv>=1.0.0 